requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-Admin==1.6.1
Flask-CORS==4.0.0
Flask-Session==0.5.0
WTForms==3.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
email-validator==2.1.0
stripe==7.8.0
requests==2.31.0
beautifulsoup4==4.12.2
python-slugify==8.0.1
celery==5.3.4